packages/airflow/requirements/requirements.txt
apache-airflow[ssh]==2.2.2
apache-airflow-providers-pagerduty==2.1.0
slack-sdk==3.17.2
mysql-connector-python==8.0.29